+++ selftests/unit/test_loader.py
@@ -4,6 +4,7 @@ import tempfile
 import unittest.mock
 
 from avocado.core import test
+from avocado.core.test_id import TestID
 from avocado.core import loader
 from avocado.utils import script
 
@@ -178,7 +179,7 @@ class LoaderTest(unittest.TestCase):
         test_class, test_parameters = (
             self.loader.discover(simple_test.path, loader.DiscoverMode.ALL)[0])
         self.assertTrue(test_class == test.SimpleTest, test_class)
-        test_parameters['name'] = test.TestID(0, test_parameters['name'])
+        test_parameters['name'] = TestID(0, test_parameters['name'])
         test_parameters['base_logdir'] = self.tmpdir.name
         tc = test_class(**test_parameters)
         tc.run_avocado()
@@ -224,7 +225,7 @@ class LoaderTest(unittest.TestCase):
         test_class, test_parameters = (
             self.loader.discover(avocado_not_a_test.path, loader.DiscoverMode.ALL)[0])
         self.assertTrue(test_class == test.SimpleTest, test_class)
-        test_parameters['name'] = test.TestID(0, test_parameters['name'])
+        test_parameters['name'] = TestID(0, test_parameters['name'])
         test_parameters['base_logdir'] = self.tmpdir.name
         tc = test_class(**test_parameters)
         # The test can't be executed (no shebang), raising an OSError
@@ -240,7 +241,7 @@ class LoaderTest(unittest.TestCase):
             test_class, test_parameters = (
                 self.loader.discover(avocado_simple_test.path, loader.DiscoverMode.ALL)[0])
             self.assertTrue(test_class == test.SimpleTest)
-            test_parameters['name'] = test.TestID(0, test_parameters['name'])
+            test_parameters['name'] = TestID(0, test_parameters['name'])
             test_parameters['base_logdir'] = self.tmpdir.name
             tc = test_class(**test_parameters)
             tc.run_avocado()
